add post build function to generate docs on build
